Refactor Common module - Common 모듈의 스프링 의존성 제거
