+++ account/src/main/java/kr/flab/movieon/account/domain/exception/EntityNotFoundException.java
@@ -1,4 +1,4 @@
-package kr.flab.movieon.common.exception;
+package kr.flab.movieon.account.domain.exception;
 
 import kr.flab.movieon.common.ErrorCode;
