+++ account/src/main/java/kr/flab/movieon/account/infrastructure/security/auth/JwtAuthenticationProvider.java
@@ -3,8 +3,8 @@ package kr.flab.movieon.account.infrastructure.security.auth;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
 import kr.flab.movieon.account.domain.Account;
-import kr.flab.movieon.account.domain.AccountNotFoundException;
 import kr.flab.movieon.account.domain.AccountRepository;
+import kr.flab.movieon.account.domain.exception.AccountNotFoundException;
 import kr.flab.movieon.account.infrastructure.config.SecurityAppProperties;
 import kr.flab.movieon.account.infrastructure.security.domain.AccountContext;
 import kr.flab.movieon.account.infrastructure.security.jwt.JwtRawToken;
