+++ common/build.gradle
@@ -1,22 +1,14 @@
 plugins {
     id "io.freefair.lombok" version "5.3.0"
-    id 'org.springframework.boot'
-    id 'io.spring.dependency-management'
 }
 
 dependencies {
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.0'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.0'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.springframework.boot:spring-boot-starter-security'
 }
 
 test {
     useJUnitPlatform()
-}
-
-bootJar {
-    enabled = false
 }
 
 jar {
